backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
pandas==2.1.4
numpy==1.26.3
scikit-learn==1.4.0
xgboost==2.0.3
shap==0.44.0
joblib==1.3.2
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0
aiofiles==23.2.1
httpx==0.26.0
openpyxl==3.1.2
transformers==4.37.2
torch==2.1.2
sentencepiece==0.1.99
aiosqlite==0.19.0
greenlet==3.0.3
